requirements.txt
sentence-transformers>=5.1.0
faiss-cpu>=1.12.0
numpy>=1.25.0,<2.0.0
google-genai>=1.46.0
python-dotenv>=1.1.1
sqlparse>=0.5.3
streamlit>=1.32.2
pandas>=2.3.0